+++ packages/vue-adapter/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @daltonr/pathwrite-vue
 
+## 0.14.1
+
+### Patch Changes
+
+- @daltonr/pathwrite-core@0.14.1
+
 ## 0.14.0
 
 ### Minor Changes